Revert "Add CONFIG to nstore-from procedure to allow pagination."
This reverts commit 01a446a6a7c87fa943e6ce2028e553ee2b503f88.

In fact, it is not really possible to predicte the order of the
returned tuples because the actual "subspace" (or index if you prefer)
that is used to bind the pattern is not known hence the order is not
known.

In theory, it could be possible to somehow force the use of a certain
index to force the order but that would lead to manyfolds subspaces /
indices that will require a lot of disk space.

Instead, one must create indices using okvs to do pagination.