Draft #2 (#2)
* remove 'nstore-close' and 'nstore-everything'

- 'nstore-close' not needed since okvs must manage the database

- 'nstore-everything' no need to specify it

* some typos

- nstore-engine: s/range/prefix/

- nstore: add prefix parameter

* Change from rest parameters to actual parameter.

This is done so to be consistent and allow to easily pass an optional
CONFIG to nstore-from procedure.

* Add CONFIG to nstore-from procedure to allow pagination.

* Revert "Add CONFIG to nstore-from procedure to allow pagination."

This reverts commit 01a446a6a7c87fa943e6ce2028e553ee2b503f88.

In fact, it is not really possible to predicte the order of the
returned tuples because the actual "subspace" (or index if you prefer)
that is used to bind the pattern is not known hence the order is not
known.

In theory, it could be possible to somehow force the use of a certain
index to force the order but that would lead to manyfolds subspaces /
indices that will require a lot of disk space.

Instead, one must create indices using okvs to do pagination.

* s/some/transaction/g

If we force the use of SRFI-167 transactional in those procedures this
will make application code that must rely on other transactional
semantic more difficult maybe impossible.

* Add srfi/pack.scm and replace srfi/nstore.scm with debugged version

* Revert "s/some/transaction/g"

This reverts commit 7509a86084bfccf44d16520761cd8e621a633977.

Transaction can be managed by application anyway, since the procedure
accept a transaction object. It only incurs a small performance
penality.

* remove paragraph about ordering bindings

The actual subspace used to do the query depends on the variables
that are passed, which is not known by the application. Otherwise
said it is not possible to predict the order of the bindings.

* add CONFIG to nstore-from

* Remove 'some' from 'nstore-from' and 'nstore-where'

'nstore-from' and 'nstore-select' results must be turned into a list
or something before the transaction is closed. If the transaction is
commited, it will invalidated the transaction which will in turn
invalidate the generator.

* update sample implementation with tests